+++ src/app/mod.rs
@@ -20,3 +20,4 @@ pub(crate) use keybindings::SectionKeybindings;
 pub(crate) use keybindings::TisqKeyboundAction;
 pub(crate) use model::EditorId;
 pub(crate) use user_event::TisqEvent;
+pub(crate) use snippets::Snippet;