+++ src/components/mod.rs
@@ -1,17 +1,14 @@
 use super::Msg;
 
-
-
-
-
 // -- modules
 mod add_server_form;
 mod browser;
 mod editor;
-mod execute_result_table;
 mod error_result;
+mod execute_result_table;
 mod global_keys;
 mod label;
+mod snippets_table;
 mod tabs;
 
 // -- export
@@ -21,8 +18,9 @@ pub(crate) use add_server_form::FormSubmitListener;
 pub(crate) use browser::BrowserTree;
 pub(crate) use browser::SentTree;
 pub use editor::Editor;
-pub(crate) use execute_result_table::ExecuteResultTable;
 pub(crate) use error_result::ErrorResult;
+pub(crate) use execute_result_table::ExecuteResultTable;
+pub(crate) use snippets_table::SnippetsTable;
 
 pub use global_keys::GlobalListener;
 pub use label::Label;